% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\name{LHD}
\alias{LHD}
\title{Life Hist Data creator}
\usage{
LHD(genecap_file, low = FALSE, high = FALSE, moyenne = FALSE,
  age.supp = 8)
}
\arguments{
\item{genecap_file}{The genetic-catch names links file}

\item{low}{Age is calculated with the a lower threshold with low = TRUE}

\item{high}{Age is calculated with the a upper threshold with high = TRUE}

\item{moyenne}{The mean age is calculated whith moyenne = TRUE}

\item{age.supp}{The age attributed to the 6 or more years old individuals}
}
\value{
A LifeHistData file to be used to by PediCre
}
\description{
Creates a LifeHistData taking in parameters the link file between genetic individuals names
and caught individuals names and the differents way to calculate the individuals age
and transforms the F/M sex in 1/2 sex .
}
