% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\name{LogDB}
\alias{LogDB}
\title{LogDB}
\usage{
LogDB(dbn, hst, usr, pw)
}
\arguments{
\item{dbn}{Database name}

\item{hst}{Host server url}

\item{usr}{User}

\item{pw}{Password}
}
\value{
Access to INRA database
}
\description{
Allows log-in to INRA database to pick deer data needed to build the Life Hist Data.
}
