% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pedigree.R
\name{PediCre}
\alias{PediCre}
\title{Pedigree Creator}
\usage{
PediCre(raw_file, genecap_file, LHData, MS)
}
\arguments{
\item{raw_file}{The raw file name}

\item{genecap_file}{The genetic-catch names links file}

\item{LHData}{LifeHistData file}

\item{MS1}{First run maximum number of sibship clustering iterations}
}
\value{
The population's SNP pedigree
}
\description{
Creates a Pedigree taking in paramater a raw file containing a 0/1/2 genome, the link file between genetic individuals names
and caught individuals names.
It also take the Maximum Siblings number Iterations number to relatives assignment as MS1 to the Parent-Offspring assignment
and MS2 to the others relatives assignement.
}
